+++ pkg/providers/apisix/translation/apisix_cluster_config.go
@@ -15,9 +15,9 @@
 package translation
 
 import (
-	"github.com/apache/apisix-ingress-controller/pkg/id"
-	configv2 "github.com/apache/apisix-ingress-controller/pkg/kube/apisix/apis/config/v2"
-	apisixv1 "github.com/apache/apisix-ingress-controller/pkg/types/apisix/v1"
+	"github.com/api7/api7-ingress-controller/pkg/id"
+	configv2 "github.com/api7/api7-ingress-controller/pkg/kube/apisix/apis/config/v2"
+	apisixv1 "github.com/api7/api7-ingress-controller/pkg/types/apisix/v1"
 )
 
 type prometheusPluginConfig struct {