+++ pkg/providers/k8s/endpoint/base.go
@@ -24,14 +24,14 @@ import (
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	listerscorev1 "k8s.io/client-go/listers/core/v1"
 
-	"github.com/apache/apisix-ingress-controller/pkg/config"
-	"github.com/apache/apisix-ingress-controller/pkg/kube"
-	configv2 "github.com/apache/apisix-ingress-controller/pkg/kube/apisix/apis/config/v2"
-	"github.com/apache/apisix-ingress-controller/pkg/log"
-	"github.com/apache/apisix-ingress-controller/pkg/providers/translation"
-	providertypes "github.com/apache/apisix-ingress-controller/pkg/providers/types"
-	"github.com/apache/apisix-ingress-controller/pkg/providers/utils"
-	apisixv1 "github.com/apache/apisix-ingress-controller/pkg/types/apisix/v1"
+	"github.com/api7/api7-ingress-controller/pkg/config"
+	"github.com/api7/api7-ingress-controller/pkg/kube"
+	configv2 "github.com/api7/api7-ingress-controller/pkg/kube/apisix/apis/config/v2"
+	"github.com/api7/api7-ingress-controller/pkg/log"
+	"github.com/api7/api7-ingress-controller/pkg/providers/translation"
+	providertypes "github.com/api7/api7-ingress-controller/pkg/providers/types"
+	"github.com/api7/api7-ingress-controller/pkg/providers/utils"
+	apisixv1 "github.com/api7/api7-ingress-controller/pkg/types/apisix/v1"
 )
 
 type baseEndpointController struct {