+++ test/e2e/suite-gateway/gateway_tcproute.go
@@ -24,10 +24,10 @@ import (
 	"github.com/onsi/ginkgo/v2"
 	"github.com/stretchr/testify/assert"
 
-	"github.com/apache/apisix-ingress-controller/test/e2e/scaffold"
+	"github.com/api7/api7-ingress-controller/test/e2e/scaffold"
 )
 
-var _ = ginkgo.Describe("suite-gateway: TCP Route", func() {
+var _ = ginkgo.PDescribe("suite-gateway: TCP Route", func() {
 	s := scaffold.NewDefaultScaffold()
 	ginkgo.It("create TCPRoute", func() {
 		backendSvc, backendPorts := s.DefaultHTTPBackend()