+++ test/e2e/suite-gateway/gateway_udproute.go
@@ -23,10 +23,10 @@ import (
 	"github.com/onsi/ginkgo/v2"
 	"github.com/stretchr/testify/assert"
 
-	"github.com/apache/apisix-ingress-controller/test/e2e/scaffold"
+	"github.com/api7/api7-ingress-controller/test/e2e/scaffold"
 )
 
-var _ = ginkgo.Describe("suite-gateway: UDP Route", func() {
+var _ = ginkgo.PDescribe("suite-gateway: UDP Route", func() {
 	s := scaffold.NewDefaultScaffold()
 
 	ginkgo.It("UDPRoute", func() {